Remove unnecessary cast
git-svn-id: https://svn.code.sf.net/p/sidplay-residfp/code/trunk/libsidplayfp@1530 75748be7-0387-4326-94c7-f41ce0883d2c

+++ sidplayfp/sidtune/SidTuneBase.cpp
@@ -184,7 +184,7 @@ void SidTuneBase::loadFile(const char* fileName, buffer_t& bufferRef)
     }
 
     inFile.seekg(0, inFile.end);
-    const int fileLen = (size_t)inFile.tellg();
+    const int fileLen = inFile.tellg();
 
     if (fileLen <= 0)
     {
